Move client parsers and commands into client subpackage
