+++ src/main/java/seedu/address/logic/parser/client/DeleteCommandParser.java
@@ -1,10 +1,12 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.client;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.exception.ParseException;
-import seedu.address.logic.commands.DeleteCommand;
+import seedu.address.logic.commands.client.DeleteCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.util.ParserUtil;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
