+++ src/main/java/seedu/address/logic/parser/client/EditCommandParser.java
@@ -1,12 +1,12 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.client;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
+import static seedu.address.logic.parser.util.CliSyntax.PREFIX_ADDRESS;
+import static seedu.address.logic.parser.util.CliSyntax.PREFIX_EMAIL;
+import static seedu.address.logic.parser.util.CliSyntax.PREFIX_NAME;
+import static seedu.address.logic.parser.util.CliSyntax.PREFIX_PHONE;
+import static seedu.address.logic.parser.util.CliSyntax.PREFIX_TAG;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -15,8 +15,12 @@ import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.exception.ParseException;
-import seedu.address.logic.commands.EditCommand;
+import seedu.address.logic.commands.client.EditCommand;
 import seedu.address.logic.descriptors.PersonDescriptor;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.util.ArgumentMultimap;
+import seedu.address.logic.parser.util.ArgumentTokenizer;
+import seedu.address.logic.parser.util.ParserUtil;
 import seedu.address.model.tag.Tag;
 
 /**
