+++ src/main/java/seedu/address/logic/parser/client/FindCommandParser.java
@@ -1,11 +1,12 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.client;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import java.util.Arrays;
 
 import seedu.address.exception.ParseException;
-import seedu.address.logic.commands.FindCommand;
+import seedu.address.logic.commands.client.FindCommand;
+import seedu.address.logic.parser.Parser;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 
 /**
