+++ src/main/java/seedu/address/logic/parser/util/ParserUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.util;
 
 import static java.util.Objects.requireNonNull;
