+++ src/test/java/seedu/address/logic/commands/EditCommandTest.java
@@ -19,14 +19,16 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.exception.NotFoundException;
+import seedu.address.logic.commands.client.EditCommand;
+import seedu.address.logic.commands.client.ListCommand;
 import seedu.address.logic.descriptors.PersonDescriptor;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Person;
-import seedu.address.testutil.PersonDescriptorBuilder;
 import seedu.address.testutil.PersonBuilder;
+import seedu.address.testutil.PersonDescriptorBuilder;
 
 /**
  * Contains integration tests (interaction with the Model) and unit tests for EditCommand.
