+++ src/test/java/seedu/address/logic/commands/ExitCommandTest.java
@@ -1,10 +1,11 @@
 package seedu.address.logic.commands;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.logic.commands.ExitCommand.MESSAGE_EXIT_ACKNOWLEDGEMENT;
+import static seedu.address.logic.commands.general.ExitCommand.MESSAGE_EXIT_ACKNOWLEDGEMENT;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.general.ExitCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
